Reformatação da Classe Util e separação de metodos da MainActivity em classes Controller e ProductDB. Criação da funcionalidade de adicionar a quantidade desejada de imagens.

+++ app/src/main/java/br/com/bg7/appvistoria/database/ProductDB.java
@@ -1,5 +1,7 @@
 package br.com.bg7.appvistoria.database;
 
+import com.orm.SugarRecord;
+
 import org.json.JSONObject;
 
 import java.util.List;
@@ -16,7 +18,6 @@ import br.com.bg7.appvistoria.vo.PropertyList;
 
 public class ProductDB {
 
-
     /**
      * Save product Item based on JSON String
      * @param json
@@ -25,17 +26,11 @@ public class ProductDB {
         try {
             Product product = Product.fromJson(new JSONObject(json));
             if (product != null) {
-                Util.Log.i("Sucesso! " + product.getProductYourRef());
-
                 Properties properties = product.getProperties();
-
                 product.save();             // Save product in database
                 long productId = product.getId();
                 properties.setProductId(productId);
-
                 List<PropertyList> propertyGroupList = properties.getPropertyGroupList();
-
-
                 properties.save();          // Save properties in database
                 long propertiesId = properties.getId();
 
@@ -61,30 +56,30 @@ public class ProductDB {
         return null;
     }
 
-
-
     /**
      * Delete product Item based on Product object
-     * @param json
+     * @param Product obj
      */
     public boolean delete(Product obj) {
 
         try {
             Properties properties = obj.getProperties();
-
             List<PropertyList>  listProperties =  properties.getPropertyGroupList();
             for(PropertyList list: listProperties ) {
-
                 List<Property> propertys = list.getPropertyList();
-
                 for(Property property: propertys) {
-                    property.delete();
+                    if(property.getId() != null && property.getId() > 0)
+                        property.delete();
                 }
-                list.delete();
-
+                if(list.getId() != null && list.getId() > 0)
+                    list.delete();
             }
-            properties.delete();
-            obj.delete();
+            if(properties.getId() != null && properties.getId() > 0)
+                properties.delete();
+
+            if(obj.getId() != null && obj.getId() > 0)
+                obj.delete();
+
             } catch (Exception ex) {
                 ex.printStackTrace();
                 return false;
@@ -99,23 +94,15 @@ public class ProductDB {
     public List<Product> getAll() {
         List<Product> items = Product.listAll(Product.class);
         if(items != null && items.size() > 0) {
-
             for(Product product: items) {
-
                 List<Properties> properties = Properties.find(Properties.class, "product_id = ?", String.valueOf(product.getId()));
-
                 if (properties != null && properties.size() > 0) {
                     product.setProperties(properties.get(0));
-
                     List<PropertyList> propertyGroupList = PropertyList.find(PropertyList.class, "properties_id = ?", String.valueOf(product.getProperties().getId()));
-
                     if (propertyGroupList != null && propertyGroupList.size() > 0) {
                         product.getProperties().setPropertyGroupList(propertyGroupList);
-
                         for (PropertyList propertyList : product.getProperties().getPropertyGroupList()) {
-
                             List<Property> propertys = Property.find(Property.class, "property_list_id = ?", String.valueOf(propertyList.getId()));
-
                             if (propertys != null && propertys.size() > 0) {
                                 propertyList.setPropertyList(propertys);
                             }
@@ -124,9 +111,6 @@ public class ProductDB {
                 }
             }
         }
-
         return items;
-
     }
-
 }
